+++ test/insert_test.dart
@@ -45,7 +45,7 @@ void main() {
       // Insert a first insert 2, adding 2 to the original product
       final insert2 = Insert.example(
         key: 'insert2',
-        produce: (components, previousProduct) => previousProduct + 2,
+        produce: (components, previousProduct, node) => previousProduct + 2,
         host: host,
       );
 
@@ -62,7 +62,7 @@ void main() {
       // Add insert0 before insert2, multiplying by 3
       final insert0 = Insert.example(
         key: 'insert0',
-        produce: (components, previousProduct) => previousProduct * 3,
+        produce: (components, previousProduct, node) => previousProduct * 3,
         host: host,
         index: 0,
       );
@@ -80,7 +80,7 @@ void main() {
       // The insert multiplies the previous result by 4
       final insert1 = Insert.example(
         key: 'insert1',
-        produce: (components, previousProduct) => previousProduct * 4,
+        produce: (components, previousProduct, node) => previousProduct * 4,
         host: host,
         index: 1,
       );
@@ -100,7 +100,7 @@ void main() {
       // Add insert3 after insert2 adding ten
       final insert3 = Insert.example(
         key: 'insert3',
-        produce: (components, previousProduct) => previousProduct + 10,
+        produce: (components, previousProduct, node) => previousProduct + 10,
         host: host,
         index: 3,
       );
@@ -181,7 +181,7 @@ void main() {
         final insert = NodeBluePrint(
           key: 'insert0',
           initialProduct: 0,
-          produce: (components, previousProduct) {
+          produce: (components, previousProduct, node) {
             return previousProduct * 10;
           },
         ).instantiateAsInsert(host: host);
@@ -211,7 +211,7 @@ void main() {
         final insert0 = NodeBluePrint(
           key: 'insert0',
           initialProduct: 0,
-          produce: (components, previousProduct) {
+          produce: (components, previousProduct, node) {
             return previousProduct * 10;
           },
         ).instantiateAsInsert(host: host);
@@ -219,7 +219,7 @@ void main() {
         final insert1 = NodeBluePrint(
           key: 'insert1',
           initialProduct: 0,
-          produce: (components, previousProduct) {
+          produce: (components, previousProduct, node) {
             return previousProduct * 10;
           },
         ).instantiateAsInsert(host: host);
@@ -266,7 +266,7 @@ void main() {
           key: 'insert0',
           initialProduct: 0,
           suppliers: ['factor'],
-          produce: (components, previousProduct) {
+          produce: (components, previousProduct, node) {
             final factor = components.first as int;
             return previousProduct * factor;
           },