+++ test/scope_blue_print_factory_test.dart
@@ -15,10 +15,13 @@ void main() {
     group('example', () {
       group('produce()', () {
         test('should turn components into a list of Scopes', () {
+          final scope = Scope.example();
+          final factoryInstance = factory.instantiate(scope: scope);
+
           // Test the produce function. It should return a scope blue print
           // for each row height
           final rowHeights = [10, 20, 30];
-          final rowScopes = factory.produce([rowHeights], []);
+          final rowScopes = factory.produce([rowHeights], [], factoryInstance);
           expect(rowScopes.length, 3);
 
           // Each scope should have a "RowHeight" node
@@ -26,7 +29,13 @@ void main() {
           var i = 0;
           for (final rowScope in rowScopes) {
             final rowHeightNode = rowScope.node<int>('rowHeight')!;
-            expect(rowHeightNode.produce([rowHeights], 0), rowHeights[i]);
+
+            final dummy = Node.example();
+
+            expect(
+              rowHeightNode.produce([rowHeights], 0, dummy),
+              rowHeights[i],
+            );
             i++;
           }
         });