+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## [Unreleased]
+
+### Changed
+
+- BREAKING CHANGE: produce function will require a node param
+
 ## [3.0.2] - 2025-10-21
 
 ### Fixed
@@ -483,6 +489,7 @@ Modifications can only be done via builders.
 - 'Github Actions Pipeline: Add SDK file containing flutter into
 .github/workflows to make github installing flutter and not dart SDK'
 
+[Unreleased]: https://github.com/ggsuite/supply_chain/compare/3.0.2...HEAD
 [3.0.2]: https://github.com/ggsuite/supply_chain/compare/3.0.1...3.0.2
 [3.0.1]: https://github.com/ggsuite/supply_chain/compare/3.0.0...3.0.1
 [3.0.0]: https://github.com/ggsuite/supply_chain/compare/2.0.0...3.0.0